fix[rust]: only cast physical inner types in list
